pubsub/requirements.txt
kafka-python==1.3.3
msgpack-python==0.4.8